+++ spec/integration/generate_all_endpoints_spec.rb
@@ -18,6 +18,7 @@ RSpec.describe "Generate an OpenAPI document for all endpoints", :rails_app do
     document = RailsOpenapiGenerator::Generator.new(configuration).document
 
     expect(document["paths"].keys).to contain_exactly(
+      "/api/inputs", "/api/inputs/{id}", "/api/inputs/upload",
       "/api/orphan", "/api/pages/{id}", "/api/pages/{id}/download",
       "/api/posts", "/api/reports/chained", "/api/reports/cyclic",
       "/api/reports/single", "/api/reports/via_concern",